+++ module/actor/actor.js
@@ -12,6 +12,7 @@ import {
   zeroFill,
 } from '../../lib/utilities.js'
 import { Combat } from '../combat/index.js'
+import { isActorInCombat } from '../combat/utilities.js'
 import Maneuvers, {
   MOVE_HALF,
   MOVE_NONE,